backend/requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
python-multipart==0.0.20
sqlalchemy==2.0.44
pydantic[email]==2.12.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pypdf==4.0.1
reportlab==4.0.9
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
lxml==5.1.0
google-auth==2.27.0
slowapi==0.1.9
python-magic==0.4.27
alembic==1.13.2
psycopg2-binary==2.9.9
openai==1.58.1
celery==5.4.0
redis==5.0.1
pytest==8.3.3
pytest-asyncio==0.23.8
weasyprint==60.2
pydyf==0.9.0
